+++ src/dl_azure/__init__.py
@@ -2,7 +2,7 @@
 
 from . import callbacks, datasets, executors, metrics_sources, storage, trackers
 
-__version__ = "0.0.24"
+__version__ = "0.0.25"
 
 __all__ = [
     "__version__",